refactor: colocate task domain randomization providers
