+++ src/unilab/dr/__init__.py
@@ -1,4 +1,5 @@
 from .manager import DomainRandomizationManager
+from .provider import DomainRandomizationProvider
 from .types import (
     DomainRandomizationCapabilities,
     IntervalRandomizationPlan,
@@ -9,6 +10,7 @@ from .types import (
 __all__ = [
     "DomainRandomizationCapabilities",
     "DomainRandomizationManager",
+    "DomainRandomizationProvider",
     "IntervalRandomizationPlan",
     "ResetPlan",
     "ResetRandomizationPayload",
